* vrpn_VRPNDevice.cpp (Java_vrpn_VRPNDevice_getEarliestTime_1native): 	this should prevent times from being truncated; one calculation was 	done at long precision (only 32 bits). 	(Java_vrpn_VRPNDevice_getLatestTime_1native):  same 	* java_vrpn.dsp:  path changes for a recent java installation 	(v. 1.4.2_10, the last release before 1.5).

+++ java_vrpn/vrpn_VRPNDevice.cpp
@@ -262,7 +262,7 @@ Java_vrpn_VRPNDevice_getEarliestTime_1native( JNIEnv* env, jobject jobj, jobject
             "Java method \'Date::setTime\'.\n" );
 		return false;
 	}
-	env->CallVoidMethod( jdate, jmid, (jlong) ( t.tv_sec * 1000 + t.tv_usec / 1000 ) );
+	env->CallVoidMethod( jdate, jmid, ( (jlong) t.tv_sec ) * 1000 + ( (jlong) t.tv_usec ) / 1000 );
 	return true;
 }
 
@@ -306,7 +306,7 @@ Java_vrpn_VRPNDevice_getLatestTime_1native( JNIEnv* env, jobject jobj, jobject j
             "Java method \'Date::setTime\'.\n" );
 		return false;
 	}
-	env->CallVoidMethod( jdate, jmid, (jlong) ( t.tv_sec * 1000 + t.tv_usec / 1000 ) );
+	env->CallVoidMethod( jdate, jmid, ( (jlong) t.tv_sec ) * 1000 + ( (jlong) t.tv_usec ) / 1000 );
 	return true;
 }
